requirements.txt
colorama
faicons
ffmpeg-python
openai
python-dotenv
shiny